Removed logic to get new hand from Player to GameUtilities. Returning validMove as playable if playing it results in empty hand
