+++ src/test/scala/PlayerTest.scala
@@ -17,50 +17,6 @@ class PlayerTest extends FunSpec{
     // No tests for playNextMove yet, individual methods are tested however
     // Possibly try and verify object method calls happened for unit testing
 
-    describe("Tests for getNewHand()") {
-      val currentHand = Hand(List(
-        NormalCard(SIX, Diamond),
-        NormalCard(SIX, Heart),
-        NormalCard(EIGHT, Club),
-        NormalCard(TEN, Heart),
-        NormalCard(ACE, Diamond),
-        NormalCard(ACE, Heart),
-        SpecialCard(TWO, Diamond),
-        Joker,
-      ))
-      val player = Player("Test", currentHand)
-      describe("When the movePlayed is none"){
-        it("Should return the same hand") {
-          assert(player.getNewHand(player.hand, None) == currentHand)
-        }
-      }
-
-      describe("When the movePlayed does not involve a hand"){
-        it("Should return the same hand") {
-          assert(player.getNewHand(player.hand, Some(Move(List(NormalCard(SEVEN, Diamond), NormalCard(SEVEN, Spade)))))
-                 == currentHand)
-        }
-      }
-
-      describe("When the movePlayed involves a card in the hand"){
-        describe("When the move played is a normalCard") {
-          it("Should return the hand minus the played cards") {
-            assert(player.getNewHand(player.hand, Some(Move(List(NormalCard(ACE, Diamond), NormalCard(ACE, Heart)))))
-                  == Hand(currentHand.listOfCards.slice(0,4) ++ currentHand.listOfCards.slice(6, 8)))
-          }
-        }
-      }
-
-      describe("When the move played is a specialCard") {
-        it("Should return the hand minus the played cards") {
-          assert(player.getNewHand(player.hand, Some(Move(List(SpecialCard(TWO, Diamond)))))
-                  == Hand(currentHand.listOfCards.slice(0,6) ++ currentHand.listOfCards.slice(7, 8)))
-          assert(player.getNewHand(player.hand, Some(Move(List(Joker))))
-                  == Hand(currentHand.listOfCards.slice(0,7)))
-        }
-      }
-    }
-
   }
